fix: bug fixes and logic changes
