+++ src/features/users/actions.js
@@ -2,7 +2,7 @@ export const FETCH_USERS = 'FETCH_USERS'
 export const FETCH_USERS_SUCCESS = 'FETCH_USERS_SUCCESS'
 export const FETCH_USERS_FAILURE = 'FETCH_USERS_FAILURE'
 
-export const loadUsersStart = () => ({
+export const fetchUsersStart = () => ({
     type: FETCH_USERS
 })
