+++ src/store/reducers.js
@@ -1,7 +1,9 @@
 import { combineReducers } from 'redux'
 
+import { appReducer } from '../features/app/reducers'
 import { usersReducer } from '../features/users/reducers'
 
 export const rootReducer = combineReducers({
-    users: usersReducer
+    users: usersReducer,
+    app: appReducer
 })
