+++ src/store/sagas.js
@@ -1,7 +1,9 @@
 import { all } from 'redux-saga/effects'
 
-import usersSaga from '../features/app/sagas'
+import { appSagas } from '../features/app/sagas'
+import { usersSagas } from '../features/users/sagas'
 
 export default function* rootSaga() {
-    yield all([usersSaga()])
+    yield all([appSagas()])
+    yield all([usersSagas()])
 }
