+++ src/store/store.js
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import createSagaMiddleware from 'redux-saga'
 
-import { initializeApp } from '../features/app/index'
 import { rootReducer } from './reducers'
 import rootSaga from './sagas'
 
@@ -12,7 +11,5 @@ export const store = configureStore({
     reducer: rootReducer,
     middleware: [sagaMiddleware]
 })
-
-store.dispatch(initializeApp())
 
 sagaMiddleware.run(rootSaga)
